+++ app/templates/categories/index.html
@@ -78,7 +78,7 @@
 
             {{ list_item_small("Mental capacity, mental health",
                  "Help if someone cannot make decisions about their health, day-to-day life or care. Help at mental health tribunals.",
-                 url_for("find-a-legal-advisor.search", category="mhe", secondary_category="com")) }}
+                 url_for("find-a-legal-adviser.search", category="mhe", secondary_category="com")) }}
 
             {{ list_item_small("Harassment and stalking",
                  "Get legal advice or go to court to stop someone harassing you.",
