+++ app/templates/categories/more-problems.html
@@ -27,7 +27,7 @@
 
         {{ list_item_small("Clinical negligence in babies",
              "Help if a baby has brain or nerve damage caused during pregnancy, childbirth or up to 8 weeks old.",
-             url_for("find-a-legal-advisor.search", category="med")) }}
+             url_for("find-a-legal-adviser.search", category="med")) }}
 
         {{ list_item_small("Compensation for abuse, assault or neglect",
              "Includes child abuse, sexual assault, abuse of a vulnerable adult. Claims can be against a person or an organisation.",
