+++ app/templates/find_a_legal_adviser/results.html
@@ -15,7 +15,7 @@
 {% block beforeContent%}
     {{ super() }}
     {{ govukBackLink({
-        'href': url_for('find-a-legal-advisor.search', category=category, secondary_category=secondary_category),
+        'href': url_for('find-a-legal-adviser.search', category=category, secondary_category=secondary_category),
         'text': "Back"
     }) }}
 {% endblock %}
